// packages/test-utils/src/types.ts
export interface CancelOrder {
    order: Order;
    takerAssetCancelAmount: bigint;
}

export interface BatchMatchOrder {
    leftOrders: Order[];
    rightOrders: Order[];
    leftSignatures: string[];
    rightSignatures: string[];
}

export interface MatchOrder {
    left: Order;
    right: Order;
    leftSignature: string;
    rightSignature: string;
}